Rename authnonext to denial as ext is not really a good abbreviation of non-existence, and the full term is authenticated denial of existence.
